created a file inside ../src/data named buildingFeatures to store all buildings and its features
