+++ src/components/BuildingCoordinates.tsx
@@ -1,3872 +1,11 @@
 import React, { useMemo } from 'react';
 import Mapbox from '@rnmapbox/maps';
 import * as turf from '@turf/turf';
-import { Feature, Polygon } from 'geojson';
-
-//DID NOT PUT SB BUILDING, PERMANENTLY CLOSED
-
-interface BuildingProperties {
-  height: number;
-  color: string;
-  id: string;
-}
-
-type BuildingFeature = Feature<Polygon, BuildingProperties>;
+import { buildingFeatures } from '../data/buildingFeatures.ts'
 
 interface HighlightBuildingProps {
   userCoordinates: [number, number] | null;
 }
-
-// Define the buildings
-const buildingFeatures: BuildingFeature[] = [
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'H',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [-73.57903392385268, 45.49770957233865],
-          [-73.57833781758659, 45.49737298229737],
-          [-73.57884950710131, 45.496829618863075],
-          [-73.57954413762752, 45.49716439702004],
-          [-73.57903392385268, 45.49770957233865]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'EV',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [-73.57786725246696, 45.49518863241383],
-          [-73.57723491398093, 45.49583591407978],
-          [-73.57770414928625, 45.49606457158524],
-          [-73.5780002442911, 45.49575500427824],
-          [-73.57830888569514, 45.49590099361825],
-          [-73.57833146921273, 45.49588692237316],
-          [-73.57843936823984, 45.49593265390695],
-          [-73.57873044468579, 45.49562484478864],
-          [-73.57786725246696, 45.49518863241383],
-          [-73.57786725246696, 45.49518863241383]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'EV',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [-73.57786725246696, 45.49518863241383],
-          [-73.57723491398093, 45.49583591407978],
-          [-73.57770414928625, 45.49606457158524],
-          [-73.5780002442911, 45.49575500427824],
-          [-73.57830888569514, 45.49590099361825],
-          [-73.57833146921273, 45.49588692237316],
-          [-73.57843936823984, 45.49593265390695],
-          [-73.57873044468579, 45.49562484478864],
-          [-73.57786725246696, 45.49518863241383],
-          [-73.57786725246696, 45.49518863241383]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'MB',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [-73.57878063027987, 45.49613140975805],
-          [-73.57912189231992, 45.49577787010381],
-          [-73.57908425312434, 45.495762039917935],
-          [-73.57906417888618, 45.49578138792185],
-          [-73.57873797252478, 45.49562308587474],
-          [-73.57842431256188, 45.49595727856385],
-          [-73.57878063027987, 45.49613140975805]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'GS',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [-73.58129519382761, 45.49676898377763],
-          [-73.58102293710081,
-            45.49666377799775
-          ],
-          [
-            -73.58106831322218,
-            45.49660016510586
-          ],
-          [
-            -73.58082747073294,
-            45.49652676552563
-          ],
-          [
-            -73.58093567532948,
-            45.49640932599766
-          ],
-          [
-            -73.58141736030744,
-            45.496558572021996
-          ],
-          [
-            -73.58144877454552,
-            45.496614845010185
-          ],
-          [
-            -73.58129170335665,
-            45.49676898377763
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'EN',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.5795667584397,
-            45.49695060326539
-          ],
-          [
-            -73.57978609930527,
-            45.49671560832212
-          ],
-          [
-            -73.5797308182741,
-            45.496689358826075
-          ],
-          [
-            -73.57962382272991,
-            45.49680810644878
-          ],
-          [
-            -73.57958459103038,
-            45.49679060681481
-          ],
-          [
-            -73.57948651178098,
-            45.49689685450849
-          ],
-          [
-            -73.57952217696294,
-            45.49691685405193
-          ],
-          [
-            -73.57951147740796,
-            45.496928103792015
-          ],
-          [
-            -73.57956854169876,
-            45.49695185323591
-          ]
-
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'X',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57975617159853,
-            45.49686571383987
-          ],
-          [
-            -73.57968235955182,
-            45.496830605610455
-          ],
-          [
-            -73.57958877642119,
-            45.49692853903784
-          ],
-          [
-            -73.57966786075696,
-            45.49696087550953
-          ],
-          [
-            -73.57975617159853,
-            45.49686571383987
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'Z',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57975221738174,
-            45.49699967925099
-          ],
-          [
-            -73.57984184629643,
-            45.4969026698472
-          ],
-          [
-            -73.57981680292343,
-            45.496888811347276
-          ],
-          [
-            -73.57983525593511,
-            45.49687218114289
-          ],
-          [
-            -73.5798602993081,
-            45.49688142014594
-          ],
-          [
-            -73.57989720533142,
-            45.496841692422095
-          ],
-          [
-            -73.57984052822415,
-            45.49681397538885
-          ],
-          [
-            -73.57978385111603,
-            45.49687587674427
-          ],
-          [
-            -73.57975485352627,
-            45.496863866038865
-          ],
-          [
-            -73.57966786075696,
-            45.49696179940844
-          ],
-          [
-            -73.57975221738174,
-            45.49699967925099
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'PR',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57982734750152,
-            45.497032939579526
-          ],
-          [
-            -73.57992488484895,
-            45.496933158534944
-          ],
-          [
-            -73.57994597400514,
-            45.496934082434336
-          ],
-          [
-            -73.57993674749929,
-            45.49692761513839
-          ],
-          [
-            -73.57996706316135,
-            45.49689065914748
-          ],
-          [
-            -73.5799894703898,
-            45.49689527864763
-          ],
-          [
-            -73.58004746556934,
-            45.49682598610494
-          ],
-          [
-            -73.57995124629417,
-            45.49678163883286
-          ],
-          [
-            -73.5798919330424,
-            45.49684908362863
-          ],
-          [
-            -73.57994465593288,
-            45.49687495284397
-          ],
-          [
-            -73.57990774990954,
-            45.496915604443984
-          ],
-          [
-            -73.57985502701905,
-            45.496892506947574
-          ],
-          [
-            -73.57974562702043,
-            45.49700799433495
-          ],
-          [
-            -73.57979571376639,
-            45.497032015681754
-          ],
-          [
-            -73.57982734750152,
-            45.497032939579526
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'V',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57992752099346,
-            45.49693223463558
-          ],
-          [
-            -73.57981153063439,
-            45.49705141753131
-          ],
-          [
-            -73.5798985234037,
-            45.49708929731361
-          ],
-          [
-            -73.58001583183504,
-            45.496968266700435
-          ],
-          [
-            -73.57992752099346,
-            45.49693223463558
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'ER',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57996725388098,
-            45.49667784665681
-          ],
-          [
-            -73.58038561998235,
-            45.49623703003519
-          ],
-          [
-            -73.58008716773145,
-            45.49614737199511
-          ],
-          [
-            -73.58007117921808,
-            45.49616044713483
-          ],
-          [
-            -73.58000456041228,
-            45.496132428974505
-          ],
-          [
-            -73.5796368246042,
-            45.49651721049122
-          ],
-          [
-            -73.57996725388098,
-            45.49667784665681
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'FA',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57949803496372,
-            45.49688257830897
-          ],
-          [
-            -73.57959734407163,
-            45.496778536672394
-          ],
-          [
-            -73.57951618824208,
-            45.49673886606968
-          ],
-          [
-            -73.57941687913421,
-            45.496842159278714
-          ],
-          [
-            -73.57949803496372,
-            45.49688257830897
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'R',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57941581129435,
-            45.49684365628039
-          ],
-          [
-            -73.57951298472251,
-            45.49674036307405
-          ],
-          [
-            -73.57942435401324,
-            45.49670144094716
-          ],
-          [
-            -73.5793325197851,
-            45.49680174021947
-          ],
-          [
-            -73.57941581129435,
-            45.49684365628039
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'RR',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57932990120985,
-            45.496801545309125
-          ],
-          [
-            -73.57946872039196,
-            45.49665334184317
-          ],
-          [
-            -73.57938115752322,
-            45.49661367115223
-          ],
-          [
-            -73.57924340618032,
-            45.49676412022828
-          ],
-          [
-            -73.57932990120985,
-            45.496801545309125
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'T',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.5793395117686,
-            45.496656335856414
-          ],
-          [
-            -73.5792668986573,
-            45.496620407686606
-          ],
-          [
-            -73.57917826794872,
-            45.496716964590945
-          ],
-          [
-            -73.57925408457895,
-            45.49675364120114
-          ],
-          [
-            -73.5793395117686,
-            45.496656335856414
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'P',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57917130013423,
-            45.49672137506647
-          ],
-          [
-            -73.57926467011498,
-            45.49662178065958
-          ],
-          [
-            -73.57919159795615,
-            45.49658763396525
-          ],
-          [
-            -73.57910093435166,
-            45.49668438287878
-          ],
-          [
-            -73.57917130013423,
-            45.49672137506647
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'Q',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57909687478727,
-            45.496682485842825
-          ],
-          [
-            -73.57919565752054,
-            45.496577200248964
-          ],
-          [
-            -73.5791293513023,
-            45.49655064168968
-          ],
-          [
-            -73.57920783621364,
-            45.496466223328184
-          ],
-          [
-            -73.57915100231236,
-            45.49643681915026
-          ],
-          [
-            -73.57897102829149,
-            45.496623677697556
-          ],
-          [
-            -73.57909687478727,
-            45.496682485842825
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'LS',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57954795154755,
-            45.49653305219917
-          ],
-          [
-            -73.57966339278748,
-            45.49641167384266
-          ],
-          [
-            -73.57956617911175,
-            45.49636269653729
-          ],
-          [
-            -73.57966035486012,
-            45.49626900070393
-          ],
-          [
-            -73.57945681372658,
-            45.496173175258576
-          ],
-          [
-            -73.57925327259254,
-            45.496392508815205
-          ],
-          [
-            -73.57954795154755,
-            45.49653305219917
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'CI',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57984081627106,
-            45.497582838500165
-          ],
-          [
-            -73.58001589085626,
-            45.49740612622293
-          ],
-          [
-            -73.5799727057919,
-            45.49738649149123
-          ],
-          [
-            -73.5799727057919,
-            45.49737094732379
-          ],
-          [
-            -73.57994469385828,
-            45.49736194806695
-          ],
-          [
-            -73.57992018341633,
-            45.49736112995265
-          ],
-          [
-            -73.57975561330551,
-            45.497539478594604
-          ],
-          [
-            -73.57984081627106,
-            45.497582838500165
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'S',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.5797987983706,
-            45.497491209981206
-          ],
-          [
-            -73.57991784908855,
-            45.497364402409886
-          ],
-          [
-            -73.57988516849929,
-            45.49734558577819
-          ],
-          [
-            -73.57988283417149,
-            45.497333314058494
-          ],
-          [
-            -73.57985248791006,
-            45.497321042335614
-          ],
-          [
-            -73.57983614761544,
-            45.497321042335614
-          ],
-          [
-            -73.57974044017479,
-            45.497420852267226
-          ],
-          [
-            -73.5797602819611,
-            45.49743230585457
-          ],
-          [
-            -73.57973693868307,
-            45.49745930358691
-          ],
-          [
-            -73.5797987983706,
-            45.497491209981206
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'M',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57980229986231,
-            45.49729159019097
-          ],
-          [
-            -73.57978479240379,
-            45.497282590921486
-          ],
-          [
-            -73.57976144912502,
-            45.49728504526783
-          ],
-          [
-            -73.5796575715378,
-            45.49738812771915
-          ],
-          [
-            -73.57973110286359,
-            45.4974265790612
-          ],
-          [
-            -73.57983264612375,
-            45.497321042335614
-          ],
-          [
-            -73.5798104700096,
-            45.497308770610545
-          ],
-          [
-            -73.57980229986231,
-            45.49729159019097
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'MU',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.5795346226395,
-            45.49795918531515
-          ],
-          [
-            -73.5797531197677,
-            45.49774783486606
-          ],
-          [
-            -73.57967737409678,
-            45.49770903614669
-          ],
-          [
-            -73.57946616020611,
-            45.49792549181848
-          ],
-          [
-            -73.5795346226395,
-            45.49795918531515
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'B',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57946324691109,
-            45.49792344978772
-          ],
-          [
-            -73.57963076137571,
-            45.49775396097718
-          ],
-          [
-            -73.57956229894232,
-            45.49771414124288
-          ],
-          [
-            -73.57939041453515,
-            45.49788771423755
-          ],
-          [
-            -73.57946324691109,
-            45.49792344978772
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'K',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57938750124013,
-            45.49788567220543
-          ],
-          [
-            -73.57964095790834,
-            45.49763756475093
-          ],
-          [
-            -73.57956229894232,
-            45.49759263982756
-          ],
-          [
-            -73.57948218332882,
-            45.49768044759955
-          ],
-          [
-            -73.5795287960494,
-            45.49770597308881
-          ],
-          [
-            -73.57948363997635,
-            45.49774885588462
-          ],
-          [
-            -73.57946179026357,
-            45.49774579282888
-          ],
-          [
-            -73.57943557060824,
-            45.49777029727031
-          ],
-          [
-            -73.57940643765787,
-            45.4977560030141
-          ],
-          [
-            -73.57931612551174,
-            45.49785095764804
-          ],
-          [
-            -73.57938750124013,
-            45.49788567220543
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'D',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57945450702599,
-            45.49770903614669
-          ],
-          [
-            -73.57941080760041,
-            45.497686573718
-          ],
-          [
-            -73.57937147811742,
-            45.497728435509714
-          ],
-          [
-            -73.57933943187204,
-            45.49771924633859
-          ],
-          [
-            -73.57923746654572,
-            45.49781726408662
-          ],
-          [
-            -73.57931758215925,
-            45.497851978664755
-          ],
-          [
-            -73.57945450702599,
-            45.49770903614669
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'MI',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57928116597128,
-            45.49769167881624
-          ],
-          [
-            -73.57925640296348,
-            45.49768044759955
-          ],
-          [
-            -73.57915735093222,
-            45.49777642337898
-          ],
-          [
-            -73.57924329313583,
-            45.49781726408662
-          ],
-          [
-            -73.57940206771532,
-            45.49764266985426
-          ],
-          [
-            -73.57935545499474,
-            45.497620207398455
-          ],
-          [
-            -73.57928116597128,
-            45.49769167881624
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'LB',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57803945129312,
-            45.49725896793379
-          ],
-          [
-            -73.57831946061094,
-            45.496987580661596
-          ],
-          [
-            -73.5782710639387,
-            45.496960926484405
-          ],
-          [
-            -73.57831254680067,
-            45.496910041200834
-          ],
-          [
-            -73.57838168490413,
-            45.496924579858046
-          ],
-          [
-            -73.57857872849799,
-            45.49672588455314
-          ],
-          [
-            -73.57853033182575,
-            45.49668953777575
-          ],
-          [
-            -73.57856144397212,
-            45.496655614095175
-          ],
-          [
-            -73.57767647625134,
-            45.4962291432245
-          ],
-          [
-            -73.57743449289019,
-            45.49647630287225
-          ],
-          [
-            -73.57768339006161,
-            45.496599882289985
-          ],
-          [
-            -73.57764536410498,
-            45.49664834473202
-          ],
-          [
-            -73.57755548457077,
-            45.49660715165882
-          ],
-          [
-            -73.5772823890632,
-            45.496890656319124
-          ],
-          [
-            -73.57803945129312,
-            45.49725896793379
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'LD',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57744270018848,
-            45.496573486841584
-          ],
-          [
-            -73.57736495903498,
-            45.49653966370772
-          ],
-          [
-            -73.57709688609272,
-            45.496797094827514
-          ],
-          [
-            -73.5771102897399,
-            45.49681024821055
-          ],
-          [
-            -73.57709152463353,
-            45.49682528064517
-          ],
-          [
-            -73.57714781995188,
-            45.49684219212887
-          ],
-          [
-            -73.57717462724622,
-            45.49681964348247
-          ],
-          [
-            -73.57720143454061,
-            45.49682715969877
-          ],
-          [
-            -73.57744270018848,
-            45.496573486841584
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'FG',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57719607023849,
-            45.49463567587674
-          ],
-          [
-            -73.57727739323904,
-            45.494679809130616
-          ],
-          [
-            -73.57775746127693,
-            45.49488944161362
-          ],
-          [
-            -73.57782304434214,
-            45.49483795300608
-          ],
-          [
-            -73.57784665424589,
-            45.4948489862829
-          ],
-          [
-            -73.57796208044094,
-            45.49475336447665
-          ],
-          [
-            -73.57794371718255,
-            45.49473681453247
-          ],
-          [
-            -73.57904026603407,
-            45.49380265422505
-          ],
-          [
-            -73.57870710406195,
-            45.49361508473629
-          ],
-          [
-            -73.57774959130923,
-            45.49444075359264
-          ],
-          [
-            -73.5776866315664,
-            45.4944058146211
-          ],
-          [
-            -73.5776341651142,
-            45.49440949240872
-          ],
-          [
-            -73.5775895686297,
-            45.4944370758071
-          ],
-          [
-            -73.57748725904841,
-            45.494383747891106
-          ],
-          [
-            -73.57719607023849,
-            45.49463567587674
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'GA',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57786721845049,
-            45.49405691094759
-          ],
-          [
-            -73.57763206939435,
-            45.49428995382627
-          ],
-          [
-            -73.57774356248171,
-            45.494351056372636
-          ],
-          [
-            -73.57797263009631,
-            45.4941251187191
-          ],
-          [
-            -73.57801925447797,
-            45.494147854624856
-          ],
-          [
-            -73.57836184232731,
-            45.49384944513429
-          ],
-          [
-            -73.57826453926928,
-            45.4937940260547
-          ],
-          [
-            -73.57792195142049,
-            45.49408106787624
-          ],
-          [
-            -73.57786721845049,
-            45.49405691094759
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'GN',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57689525492123,
-            45.493435314569325
-          ],
-          [
-            -73.57696458035447,
-            45.493357560110155
-          ],
-          [
-            -73.57699693222331,
-            45.493376998735016
-          ],
-          [
-            -73.57716793495912,
-            45.493205290649996
-          ],
-          [
-            -73.57713558309028,
-            45.49318261218454
-          ],
-          [
-            -73.5771956651324,
-            45.493121056303664
-          ],
-          [
-            -73.57744061499704,
-            45.49323444866363
-          ],
-          [
-            -73.57757002247244,
-            45.49310161759047
-          ],
-          [
-            -73.57732507260782,
-            45.49300118413137
-          ],
-          [
-            -73.57739901973662,
-            45.4929396280522
-          ],
-          [
-            -73.57732507260782,
-            45.49290399029144
-          ],
-          [
-            -73.57738977634554,
-            45.49285863310871
-          ],
-          [
-            -73.57725574717458,
-            45.492803556480595
-          ],
-          [
-            -73.57714482648139,
-            45.492910469886
-          ],
-          [
-            -73.57696458035447,
-            45.49281327588946
-          ],
-          [
-            -73.5770246623971,
-            45.492716081725206
-          ],
-          [
-            -73.57692298509451,
-            45.492683683633174
-          ],
-          [
-            -73.57686290305237,
-            45.49275171960494
-          ],
-          [
-            -73.57653938436339,
-            45.49258972906091
-          ],
-          [
-            -73.57638686841027,
-            45.49273876037856
-          ],
-          [
-            -73.57671963049033,
-            45.4929007504939
-          ],
-          [
-            -73.57670114370814,
-            45.49293638825671
-          ],
-          [
-            -73.57679357761913,
-            45.49297850558372
-          ],
-          [
-            -73.57683055118352,
-            45.49295258723237
-          ],
-          [
-            -73.57700617561443,
-            45.493043301409955
-          ],
-          [
-            -73.57674273896811,
-            45.493341361251
-          ],
-          [
-            -73.57669190031703,
-            45.49332516238715
-          ],
-          [
-            -73.57653938436339,
-            45.49348391105181
-          ],
-          [
-            -73.57657173623225,
-            45.493506589395935
-          ],
-          [
-            -73.57626208263044,
-            45.493814366021155
-          ],
-          [
-            -73.57579991307449,
-            45.49358434364936
-          ],
-          [
-            -73.57567512729464,
-            45.493710694082864
-          ],
-          [
-            -73.57614654024121,
-            45.49393747620002
-          ],
-          [
-            -73.57603099785246,
-            45.494057346378725
-          ],
-          [
-            -73.5760818365035,
-            45.494073545031966
-          ],
-          [
-            -73.57603099785246,
-            45.49413186014499
-          ],
-          [
-            -73.5762482175433,
-            45.49421285325806
-          ],
-          [
-            -73.57643308536576,
-            45.49407678476203
-          ],
-          [
-            -73.57709860952589,
-            45.49440723625253
-          ],
-          [
-            -73.57733893769448,
-            45.49415777795363
-          ],
-          [
-            -73.57741288482376,
-            45.49418369575102
-          ],
-          [
-            -73.57745910177924,
-            45.49413509987173
-          ],
-          [
-            -73.57760699603686,
-            45.49398283251384
-          ],
-          [
-            -73.57736666786778,
-            45.49387592114422
-          ],
-          [
-            -73.57722339530572,
-            45.49401199012493
-          ],
-          [
-            -73.57704777087484,
-            45.49419017519833
-          ],
-          [
-            -73.57690912000785,
-            45.494102702596024
-          ],
-          [
-            -73.57659946640555,
-            45.49395043515054
-          ],
-          [
-            -73.576502410799,
-            45.49392127750764
-          ],
-          [
-            -73.57665954844819,
-            45.49390831855038
-          ],
-          [
-            -73.57682130779243,
-            45.49398283251384
-          ],
-          [
-            -73.57681668609686,
-            45.49401199012493
-          ],
-          [
-            -73.57687214644348,
-            45.494054106647496
-          ],
-          [
-            -73.57689525492123,
-            45.494024949058286
-          ],
-          [
-            -73.57700617561443,
-            45.49408002449192
-          ],
-          [
-            -73.57716793495912,
-            45.49390183907067
-          ],
-          [
-            -73.57677509083693,
-            45.493723653085596
-          ],
-          [
-            -73.57672425218588,
-            45.49375929032769
-          ],
-          [
-            -73.5766503050571,
-            45.49370745433171
-          ],
-          [
-            -73.57679819931468,
-            45.49356814485543
-          ],
-          [
-            -73.57689525492123,
-            45.493616741223235
-          ],
-          [
-            -73.57683979457462,
-            45.493684776068505
-          ],
-          [
-            -73.57686290305237,
-            45.493736612085314
-          ],
-          [
-            -73.57690912000785,
-            45.49375929032769
-          ],
-          [
-            -73.57695533696335,
-            45.49373013258585
-          ],
-          [
-            -73.57713096139473,
-            45.493801407039314
-          ],
-          [
-            -73.57725112547901,
-            45.49373013258585
-          ],
-          [
-            -73.57732507260782,
-            45.49360378219595
-          ],
-          [
-            -73.5771633132636,
-            45.493529267730935
-          ],
-          [
-            -73.5771910434369,
-            45.493480671287614
-          ],
-          [
-            -73.57707550104814,
-            45.493454753167356
-          ],
-          [
-            -73.57702928409266,
-            45.493493630343266
-          ],
-          [
-            -73.57689525492123,
-            45.493435314569325
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'VA',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.57377681633824,
-            45.49617394133054
-          ],
-          [
-            -73.57428483184525,
-            45.49565251120859
-          ],
-          [
-            -73.57374830526375,
-            45.49539270276074
-          ],
-          [
-            -73.57349429750984,
-            45.49565251120859
-          ],
-          [
-            -73.57380273549627,
-            45.495781506566374
-          ],
-          [
-            -73.57354354391107,
-            45.496063115123434
-          ],
-          [
-            -73.57377681633824,
-            45.49617394133054
-          ]
-        ]
-      ]
-    },
-  },
-  //LOYOLA
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'RA',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.63790381216943,
-            45.45701541921022
-          ],
-          [
-            -73.63796150118657,
-            45.45696596120334
-          ],
-          [
-            -73.63803841987612,
-            45.456997434485004
-          ],
-          [
-            -73.63800637042229,
-            45.457037900108105
-          ],
-          [
-            -73.6383332748519,
-            45.45715929680185
-          ],
-          [
-            -73.6385383913572,
-            45.45688952601631
-          ],
-          [
-            -73.63822430670855,
-            45.45677262494124
-          ],
-          [
-            -73.63817943747328,
-            45.45684006789892
-          ],
-          [
-            -73.63807046932993,
-            45.45678611353952
-          ],
-          [
-            -73.63816020780091,
-            45.456687197080015
-          ],
-          [
-            -73.63735256156214,
-            45.45636346927208
-          ],
-          [
-            -73.63698078789679,
-            45.45665572362449
-          ],
-          [
-            -73.63790381216943,
-            45.45701541921022
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'PC',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.63786535282463,
-            45.45700193066685
-          ],
-          [
-            -73.63699360767822,
-            45.45665122741505
-          ],
-          [
-            -73.63676285161007,
-            45.456925495528594
-          ],
-          [
-            -73.6376474165384,
-            45.457276197074606
-          ],
-          [
-            -73.63786535282463,
-            45.45700193066685
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'DO',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.63594879542407,
-            45.45834627290259
-          ],
-          [
-            -73.6371025757653,
-            45.457388600930955
-          ],
-          [
-            -73.63635902843463,
-            45.4569389840662
-          ],
-          [
-            -73.63522447776577,
-            45.457919144220426
-          ],
-          [
-            -73.63594879542407,
-            45.45834627290259
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'VL',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.63827819528595,
-            45.458814730333955
-          ],
-          [
-            -73.63822566863618,
-            45.45885827244146
-          ],
-          [
-            -73.63818269228591,
-            45.458844874873535
-          ],
-          [
-            -73.63816836683628,
-            45.45887836878762
-          ],
-          [
-            -73.63812061533662,
-            45.45886162183308
-          ],
-          [
-            -73.63809673958609,
-            45.45889176634756
-          ],
-          [
-            -73.6380585383866,
-            45.458871670006715
-          ],
-          [
-            -73.63802511233658,
-            45.458898465126595
-          ],
-          [
-            -73.63798213598703,
-            45.458888416958246
-          ],
-          [
-            -73.63783888148731,
-            45.45909942812213
-          ],
-          [
-            -73.63814449108646,
-            45.45922000557596
-          ],
-          [
-            -73.63820179288635,
-            45.45914297001028
-          ],
-          [
-            -73.63866975758421,
-            45.459320486590485
-          ],
-          [
-            -73.63877003573364,
-            45.45917311437486
-          ],
-          [
-            -73.63892761568299,
-            45.459226704315995
-          ],
-          [
-            -73.63899924293318,
-            45.459153018133776
-          ],
-          [
-            -73.63894194113331,
-            45.4591329218855
-          ],
-          [
-            -73.63898491748287,
-            45.4590860306115
-          ],
-          [
-            -73.63827819528595,
-            45.458814730333955
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'VE',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.63884166298388,
-            45.45902909115526
-          ],
-          [
-            -73.63898969263298,
-            45.45882477851541
-          ],
-          [
-            -73.63843577523527,
-            45.45861041691688
-          ],
-          [
-            -73.63828297043607,
-            45.458814730333955
-          ],
-          [
-            -73.63884166298388,
-            45.45902909115526
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'PT',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.638660207284,
-            45.45933053468184
-          ],
-          [
-            -73.63914249743219,
-            45.459491303904
-          ],
-          [
-            -73.639228450132,
-            45.45936067894618
-          ],
-          [
-            -73.6391950240827,
-            45.45935063085969
-          ],
-          [
-            -73.63921889983251,
-            45.459327185318074
-          ],
-          [
-            -73.63877481088376,
-            45.4591764637473
-          ],
-          [
-            -73.638660207284,
-            45.45933053468184
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'SC',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.63920340049873,
-            45.45932031983958
-          ],
-          [
-            -73.63921602086333,
-            45.45929882183202
-          ],
-          [
-            -73.6392448674099,
-            45.459310203131366
-          ],
-          [
-            -73.63926469941086,
-            45.459284911352285
-          ],
-          [
-            -73.63930436341278,
-            45.45930135100983
-          ],
-          [
-            -73.63941073505451,
-            45.45917236279922
-          ],
-          [
-            -73.63938910014433,
-            45.45916477524841
-          ],
-          [
-            -73.6394323699647,
-            45.459112926955726
-          ],
-          [
-            -73.63910784631237,
-            45.45898393831408
-          ],
-          [
-            -73.63903212412698,
-            45.45908510590115
-          ],
-          [
-            -73.63898524848814,
-            45.45906866618091
-          ],
-          [
-            -73.63894378157697,
-            45.459133160441525
-          ],
-          [
-            -73.6389888543066,
-            45.45915086473565
-          ],
-          [
-            -73.63893476703139,
-            45.45922800480855
-          ],
-          [
-            -73.63920340049873,
-            45.45932031983958
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'SP',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.6407647736335,
-            45.4572499976447
-          ],
-          [
-            -73.6414723634341,
-            45.4575244789134
-          ],
-          [
-            -73.641169728514,
-            45.457907931472
-          ],
-          [
-            -73.6411318630395,
-            45.4578932304326
-          ],
-          [
-            -73.6410642397323,
-            45.4579788796594
-          ],
-          [
-            -73.6411144047975,
-            45.4579983354857
-          ],
-          [
-            -73.6408610995721,
-            45.4583161707973
-          ],
-          [
-            -73.6409214427859,
-            45.4583394803179
-          ],
-          [
-            -73.6410381860445,
-            45.458192750888
-          ],
-          [
-            -73.6412013573026,
-            45.4582559285046
-          ],
-          [
-            -73.6412619490213,
-            45.4581791286045
-          ],
-          [
-            -73.6413376805619,
-            45.4582085307675
-          ],
-          [
-            -73.641283454495,
-            45.4582772679952
-          ],
-          [
-            -73.6414128960257,
-            45.4583273835942
-          ],
-          [
-            -73.641925074392,
-            45.4576724722197
-          ],
-          [
-            -73.6418446166943,
-            45.4576410498015
-          ],
-          [
-            -73.6420037411084,
-            45.4574384492673
-          ],
-          [
-            -73.6414125410386,
-            45.4572092219752
-          ],
-          [
-            -73.6414323029863,
-            45.4571842470099
-          ],
-          [
-            -73.641391457363,
-            45.4571683714521
-          ],
-          [
-            -73.641382695127,
-            45.4571794744891
-          ],
-          [
-            -73.6413039848223,
-            45.4571488981991
-          ],
-          [
-            -73.6412958089382,
-            45.4571593498403
-          ],
-          [
-            -73.6409931876149,
-            45.4570419346156
-          ],
-          [
-            -73.6410125229322,
-            45.4570173396417
-          ],
-          [
-            -73.6409584909684,
-            45.4569961366528
-          ],
-          [
-            -73.6409358533229,
-            45.4570249665488
-          ],
-          [
-            -73.6408287050145,
-            45.4569834322577
-          ],
-          [
-            -73.6406556646887,
-            45.45720242332
-          ],
-          [
-            -73.6407681839922,
-            45.4572459260257
-          ],
-          [
-            -73.6407647736335,
-            45.4572499976447
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'BB',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.63916734239889,
-            45.459821094109316
-          ],
-          [
-            -73.63924847331195,
-            45.45971866326323
-          ],
-          [
-            -73.63917635694447,
-            45.45969084250751
-          ],
-          [
-            -73.63909883184984,
-            45.45979074424835
-          ],
-          [
-            -73.63916734239889,
-            45.459821094109316
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'BH',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.63909522603139,
-            45.45979074424835
-          ],
-          [
-            -73.63917635694447,
-            45.459693371667754
-          ],
-          [
-            -73.63909342312215,
-            45.45966175715728
-          ],
-          [
-            -73.63902130675523,
-            45.45976292352819
-          ],
-          [
-            -73.63909522603139,
-            45.45979074424835
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'GE',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.6407407641922,
-            45.4569453065756
-          ],
-          [
-            -73.6406692043837,
-            45.4569188764311
-          ],
-          [
-            -73.640688534775,
-            45.4568932484358
-          ],
-          [
-            -73.6406295416664,
-            45.4568714647194
-          ],
-          [
-            -73.6406110651085,
-            45.4568963330836
-          ],
-          [
-            -73.6403467909331,
-            45.4567987024899
-          ],
-          [
-            -73.6401650980098,
-            45.4570407576262
-          ],
-          [
-            -73.6404419407738,
-            45.4571430357757
-          ],
-          [
-            -73.6404514723854,
-            45.4571302758573
-          ],
-          [
-            -73.6405691920397,
-            45.4571739519857
-          ],
-          [
-            -73.6407407641922,
-            45.4569453065756
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'PS',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.64013964772174,
-            45.459956782100704
-          ],
-          [
-            -73.64032156080964,
-            45.459701591087935
-          ],
-          [
-            -73.64014755698665,
-            45.459636868546
-          ],
-          [
-            -73.64015810267317,
-            45.45960358263861
-          ],
-          [
-            -73.63966245541893,
-            45.45941496212427
-          ],
-          [
-            -73.63964136404667,
-            45.459439002028915
-          ],
-          [
-            -73.63953590718454,
-            45.45939277143404
-          ],
-          [
-            -73.63958072635072,
-            45.45933174699087
-          ],
-          [
-            -73.63944626885139,
-            45.4592762701667
-          ],
-          [
-            -73.63921426375407,
-            45.45957214592977
-          ],
-          [
-            -73.63935135767501,
-            45.459625773245904
-          ],
-          [
-            -73.63931972061619,
-            45.459662757571806
-          ],
-          [
-            -73.64013964772174,
-            45.459956782100704
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'SI',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [-73.6426807599695,
-            45.4579640331272
-          ],
-          [
-            -73.6425562352174,
-            45.457915968312
-          ],
-          [
-            -73.6425362347883,
-            45.4579413512968
-          ],
-          [
-            -73.6425111523624,
-            45.457931623775
-          ],
-          [
-            -73.6425148814004,
-            45.4579270088665
-          ],
-          [
-            -73.6424767485285,
-            45.4579124165973
-          ],
-          [
-            -73.6425210879317,
-            45.4578556503426
-          ],
-          [
-            -73.6424927571127,
-            45.4578448562178
-          ],
-          [
-            -73.6425328890628,
-            45.4577935196719
-          ],
-          [
-            -73.6425006899618,
-            45.4577809780384
-          ],
-          [
-            -73.6425335798694,
-            45.4577391980194
-          ],
-          [
-            -73.642508042354,
-            45.4577295512323
-          ],
-          [
-            -73.6425643657384,
-            45.4576575826336
-          ],
-          [
-            -73.6425578405587,
-            45.4576551502813
-          ],
-          [
-            -73.6425721419068,
-            45.4576370430182
-          ],
-          [
-            -73.6423956843836,
-            45.457569411333
-          ],
-          [
-            -73.6423820490262,
-            45.4575867313559
-          ],
-          [
-            -73.6423855932416,
-            45.4575879888619
-          ],
-          [
-            -73.6423527032251,
-            45.4576297686895
-          ],
-          [
-            -73.6423324280894,
-            45.4576219264721
-          ],
-          [
-            -73.6423180751762,
-            45.4576404686079
-          ],
-          [
-            -73.6422985510487,
-            45.4576327365838
-          ],
-          [
-            -73.6422911212593,
-            45.4576422654536
-          ],
-          [
-            -73.6423031248212,
-            45.4576468290997
-          ],
-          [
-            -73.6422705818342,
-            45.45768836492
-          ],
-          [
-            -73.6422423308534,
-            45.4576774349737
-          ],
-          [
-            -73.6422022781943,
-            45.457728635467
-          ],
-          [
-            -73.6421788342758,
-            45.4577195907836
-          ],
-          [
-            -73.6421345457242,
-            45.457775921777
-          ],
-          [
-            -73.6421023749224,
-            45.4577636791435
-          ],
-          [
-            -73.6420487145874,
-            45.4578324988567
-          ],
-          [
-            -73.6420223976498,
-            45.4578224425818
-          ],
-          [
-            -73.6419962193196,
-            45.4578559156965
-          ],
-          [
-            -73.6420527730102,
-            45.4578773409018
-          ],
-          [
-            -73.6420033712975,
-            45.4579402962029
-          ],
-          [
-            -73.6420791819524,
-            45.4579695618183
-          ],
-          [
-            -73.6420810763035,
-            45.4579944704178
-          ],
-          [
-            -73.6422394622925,
-            45.458055026585
-          ],
-          [
-            -73.6422687203591,
-            45.4580423586529
-          ],
-          [
-            -73.6424145678278,
-            45.4580985673754
-          ],
-          [
-            -73.642399144896,
-            45.4581175426731
-          ],
-          [
-            -73.6425233745908,
-            45.4581654170308
-          ],
-          [
-            -73.6426807599695,
-            45.4579640331272]
-
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'CJ',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.6400450737138,
-            45.4576216477955
-          ],
-          [
-            -73.6403151488875,
-            45.4577257361344
-          ],
-          [
-            -73.6402918671827,
-            45.4577546194322
-          ],
-          [
-            -73.6404851757882,
-            45.4578310700999
-          ],
-          [
-            -73.6406316904242,
-            45.457650659767
-          ],
-          [
-            -73.6405008133219,
-            45.4575970881555
-          ],
-          [
-            -73.640716895244,
-            45.4573339709523
-          ],
-          [
-            -73.6406388557913,
-            45.4573036400003
-          ],
-          [
-            -73.6406579523542,
-            45.457279452329
-          ],
-          [
-            -73.6403940765995,
-            45.4571762753505
-          ],
-          [
-            -73.6402077374489,
-            45.4574113088604
-          ],
-          [
-            -73.6400761821297,
-            45.4573613245086
-          ],
-          [
-            -73.640048191077,
-            45.4573110529149
-          ],
-          [
-            -73.640073144594,
-            45.457305718338
-          ],
-          [
-            -73.6400178236508,
-            45.4572141215657
-          ],
-          [
-            -73.6399909000415,
-            45.4572113501541
-          ],
-          [
-            -73.6399623791951,
-            45.4572161611527
-          ],
-          [
-            -73.6399184417727,
-            45.4572253465287
-          ],
-          [
-            -73.6398854147457,
-            45.4572294961804
-          ],
-          [
-            -73.6398604637314,
-            45.4572407305691
-          ],
-          [
-            -73.6398278558022,
-            45.4572591005164
-          ],
-          [
-            -73.6398030306958,
-            45.4572794976135
-          ],
-          [
-            -73.6397839770805,
-            45.457307083725
-          ],
-          [
-            -73.6397693061575,
-            45.4573320683351
-          ],
-          [
-            -73.6397631639899,
-            45.4573581571713
-          ],
-          [
-            -73.6397623124263,
-            45.4573863498762
-          ],
-          [
-            -73.6397651466226,
-            45.457406529222
-          ],
-          [
-            -73.6397721500353,
-            45.4574296802692
-          ],
-          [
-            -73.6397883175915,
-            45.4574469493556
-          ],
-          [
-            -73.6398202437995,
-            45.4574801001292
-          ],
-          [
-            -73.6399537185258,
-            45.4574632594502
-          ],
-          [
-            -73.6399446805309,
-            45.4574460043143
-          ],
-          [
-            -73.6400282785417,
-            45.4574352650406
-          ],
-          [
-            -73.6401513336609,
-            45.4574844444917
-          ],
-          [
-            -73.6400450737138,
-            45.4576216477955
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'HC',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.64200531770703,
-            45.459885652046324
-          ],
-          [
-            -73.64208744340056,
-            45.45977850914451
-          ],
-          [
-            -73.64212686373308,
-            45.45979002989685
-          ],
-          [
-            -73.64225662232833,
-            45.45962758707799
-          ],
-          [
-            -73.64201353027667,
-            45.45952389992746
-          ],
-          [
-            -73.6418821291677,
-            45.459701320046406
-          ],
-          [
-            -73.64190512436144,
-            45.45971053666065
-          ],
-          [
-            -73.64182956872382,
-            45.459818831765716
-          ],
-          [
-            -73.64200531770703,
-            45.459885652046324
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'JR',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.6433481748265,
-            45.4585591499338
-          ],
-          [
-            -73.6434014499521,
-            45.4584941068952
-          ],
-          [
-            -73.6433283981106,
-            45.4584646569441
-          ],
-          [
-            -73.6433412642807,
-            45.4584489935071
-          ],
-          [
-            -73.6432425098978,
-            45.4584091351866
-          ],
-          [
-            -73.6432279129406,
-            45.4584270519682
-          ],
-          [
-            -73.6431541101892,
-            45.458397491732
-          ],
-          [
-            -73.643096280091,
-            45.4584682082598
-          ],
-          [
-            -73.6431224663152,
-            45.4584788351096
-          ],
-          [
-            -73.6431130101163,
-            45.4584904267042
-          ],
-          [
-            -73.6430992598496,
-            45.4584850168532
-          ],
-          [
-            -73.643055013337,
-            45.458538847918
-          ],
-          [
-            -73.6431312624657,
-            45.458569799674
-          ],
-          [
-            -73.6431197796985,
-            45.4585834537778
-          ],
-          [
-            -73.6432203324316,
-            45.4586237236778
-          ],
-          [
-            -73.6432354360966,
-            45.4586052912001
-          ],
-          [
-            -73.6433083004756,
-            45.4586347137707
-          ],
-          [
-            -73.6433647183882,
-            45.4585657070598
-          ],
-          [
-            -73.6433481748265,
-            45.4585591499338
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'HU',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.64147939262746,
-            45.458571059038576
-          ],
-          [
-            -73.64188137843888,
-            45.45872410282257
-          ],
-          [
-            -73.64210924205402,
-            45.45843309310362
-          ],
-          [
-            -73.64169865761502,
-            45.45827250937759
-          ],
-          [
-            -73.64169220864531,
-            45.45828004852976
-          ],
-          [
-            -73.64168253518973,
-            45.458274771123484
-          ],
-          [
-            -73.64166748759243,
-            45.45829814249112
-          ],
-          [
-            -73.64168468484633,
-            45.45830417381029
-          ],
-          [
-            -73.64147939262746,
-            45.458571059038576
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'HA',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.6408687943941,
-            45.4595003354675
-          ],
-          [
-            -73.6413464049925,
-            45.4596910748883
-          ],
-          [
-            -73.6415474946086,
-            45.4594416727387
-          ],
-          [
-            -73.6414545399991,
-            45.4594044330773
-          ],
-          [
-            -73.6414628242831,
-            45.4593941442679
-          ],
-          [
-            -73.6411511972266,
-            45.4592697997904
-          ],
-          [
-            -73.6411429128496,
-            45.4592800885425
-          ],
-          [
-            -73.6410698874647,
-            45.459250935449
-          ],
-          [
-            -73.6408687943941,
-            45.4595003354675
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'HB',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.6422436438324,
-            45.4591399912472
-          ],
-          [
-            -73.6422713426249,
-            45.4591049702739
-          ],
-          [
-            -73.6419063705945,
-            45.4589627714889
-          ],
-          [
-            -73.6418792578238,
-            45.4589971408267
-          ],
-          [
-            -73.6418103176288,
-            45.4589700627892
-          ],
-          [
-            -73.641495100383,
-            45.4593683495053
-          ],
-          [
-            -73.6419795323706,
-            45.459556841626
-          ],
-          [
-            -73.642003263324,
-            45.4595268428534
-          ],
-          [
-            -73.6419418430954,
-            45.4595029331161
-          ],
-          [
-            -73.6420567168413,
-            45.4593579895672
-          ],
-          [
-            -73.6420695003405,
-            45.459362962778
-          ],
-          [
-            -73.6420883290704,
-            45.4593388819682
-          ],
-          [
-            -73.6421444864352,
-            45.4593609868397
-          ],
-          [
-            -73.6423014110025,
-            45.459162479683
-          ],
-          [
-            -73.6422436438324,
-            45.4591399912472
-          ]
-        ],
-        [
-          [
-            -73.6420270234696,
-            45.4592897678687
-          ],
-          [
-            -73.641803150824,
-            45.4592024915459
-          ],
-          [
-            -73.6419267585927,
-            45.4590461464361
-          ],
-          [
-            -73.6421508179797,
-            45.4591334496977
-          ],
-          [
-            -73.6420270234696,
-            45.4592897678687
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'CC',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.6408772869513,
-            45.4585890444863
-          ],
-          [
-            -73.6407985119765,
-            45.4586801381204
-          ],
-          [
-            -73.6410434135512,
-            45.4587656349228
-          ],
-          [
-            -73.6410337817405,
-            45.4587851652835
-          ],
-          [
-            -73.6411307292453,
-            45.4588225485177
-          ],
-          [
-            -73.641145397086,
-            45.4588024297448
-          ],
-          [
-            -73.6411619356442,
-            45.4588079540898
-          ],
-          [
-            -73.6412836990154,
-            45.4586467679607
-          ],
-          [
-            -73.6412704836223,
-            45.4586411413492
-          ],
-          [
-            -73.641376403155,
-            45.4585062439627
-          ],
-          [
-            -73.6413154218922,
-            45.4584840209662
-          ],
-          [
-            -73.6412767428744,
-            45.4585370462476
-          ],
-          [
-            -73.6412461196636,
-            45.4585235554084
-          ],
-          [
-            -73.6412288337043,
-            45.4585453547232
-          ],
-          [
-            -73.6411366941731,
-            45.4585098572153
-          ],
-          [
-            -73.6412021937158,
-            45.4584281724403
-          ],
-          [
-            -73.6410774329832,
-            45.4583815457209
-          ],
-          [
-            -73.6410079055759,
-            45.4584725211565
-          ],
-          [
-            -73.6408517473394,
-            45.458413027044
-          ],
-          [
-            -73.6408116759182,
-            45.4584611281071
-          ],
-          [
-            -73.6407504451182,
-            45.4584372457533
-          ],
-          [
-            -73.6407102295275,
-            45.4584887179595
-          ],
-          [
-            -73.6407733946252,
-            45.4585134741824
-          ],
-          [
-            -73.6407529936233,
-            45.4585395360268
-          ],
-          [
-            -73.6408772869513,
-            45.4585890444863
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'CC2',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.640685034277,
-            45.4585208608007
-          ],
-          [
-            -73.6407957736703,
-            45.4583791494107
-          ],
-          [
-            -73.6405590484041,
-            45.458288067748
-          ],
-          [
-            -73.640564372429,
-            45.4582807370996
-          ],
-          [
-            -73.6402606533331,
-            45.4581655166129
-          ],
-          [
-            -73.6402556996114,
-            45.4581718691469
-          ],
-          [
-            -73.6400146981164,
-            45.4580776515112
-          ],
-          [
-            -73.6399947518581,
-            45.4581036322558
-          ],
-          [
-            -73.6399668476527,
-            45.4580924575034
-          ],
-          [
-            -73.6399436219669,
-            45.4581219391978
-          ],
-          [
-            -73.6399074228682,
-            45.4581082196506
-          ],
-          [
-            -73.6399223691219,
-            45.4580900595336
-          ],
-          [
-            -73.6399135855623,
-            45.4580862631194
-          ],
-          [
-            -73.6400076738254,
-            45.4579642333517
-          ],
-          [
-            -73.6400664064559,
-            45.4579871594348
-          ],
-          [
-            -73.6401266871755,
-            45.4579119371099
-          ],
-          [
-            -73.6398312888425,
-            45.4577985269392
-          ],
-          [
-            -73.6397706190755,
-            45.4578764941739
-          ],
-          [
-            -73.6398149533159,
-            45.4578930300438
-          ],
-          [
-            -73.6398066969935,
-            45.4579036174965
-          ],
-          [
-            -73.6398351180342,
-            45.4579163430144
-          ],
-          [
-            -73.6397535448525,
-            45.4580219205454
-          ],
-          [
-            -73.639741129547,
-            45.4580208358951
-          ],
-          [
-            -73.6396715170233,
-            45.4579983776984
-          ],
-          [
-            -73.6396309215869,
-            45.4580654273612
-          ],
-          [
-            -73.6396882783852,
-            45.4580865293883
-          ],
-          [
-            -73.6396887773383,
-            45.4580898474012
-          ],
-          [
-            -73.6396951868686,
-            45.4580959502411
-          ],
-          [
-            -73.6396166169914,
-            45.4582019689471
-          ],
-          [
-            -73.6395811225177,
-            45.4581898275843
-          ],
-          [
-            -73.6395737199856,
-            45.4581996554648
-          ],
-          [
-            -73.6395088037371,
-            45.4581730191629
-          ],
-          [
-            -73.6394429805971,
-            45.4582552453557
-          ],
-          [
-            -73.6397634323295,
-            45.4583770539877
-          ],
-          [
-            -73.639823318343,
-            45.4582976433774
-          ],
-          [
-            -73.6397656937339,
-            45.458276649559
-          ],
-          [
-            -73.6398615938007,
-            45.4581522303785
-          ],
-          [
-            -73.6398702925016,
-            45.4581551294048
-          ],
-          [
-            -73.6398811620684,
-            45.4581418281119
-          ],
-          [
-            -73.63991569451,
-            45.4581545656232
-          ],
-          [
-            -73.6399458518892,
-            45.4581660710766
-          ],
-          [
-            -73.6399018618318,
-            45.4582187587328
-          ],
-          [
-            -73.6401444564348,
-            45.4583151280049
-          ],
-          [
-            -73.6401414600105,
-            45.4583216203185
-          ],
-          [
-            -73.640445805528,
-            45.45843855574
-          ],
-          [
-            -73.6404494963462,
-            45.4584315753495
-          ],
-          [
-            -73.640685034277,
-            45.4585208608007
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'RF',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.6408772869513,
-            45.4585890444863
-          ],
-          [
-            -73.6407985119765,
-            45.4586801381204
-          ],
-          [
-            -73.6410434135512,
-            45.4587656349228
-          ],
-          [
-            -73.6410337817405,
-            45.4587851652835
-          ],
-          [
-            -73.6411307292453,
-            45.4588225485177
-          ],
-          [
-            -73.641145397086,
-            45.4588024297448
-          ],
-          [
-            -73.6411619356442,
-            45.4588079540898
-          ],
-          [
-            -73.6412836990154,
-            45.4586467679607
-          ],
-          [
-            -73.6412704836223,
-            45.4586411413492
-          ],
-          [
-            -73.641376403155,
-            45.4585062439627
-          ],
-          [
-            -73.6413154218922,
-            45.4584840209662
-          ],
-          [
-            -73.6412767428744,
-            45.4585370462476
-          ],
-          [
-            -73.6412461196636,
-            45.4585235554084
-          ],
-          [
-            -73.6412288337043,
-            45.4585453547232
-          ],
-          [
-            -73.6411366941731,
-            45.4585098572153
-          ],
-          [
-            -73.6412021937158,
-            45.4584281724403
-          ],
-          [
-            -73.6410774329832,
-            45.4583815457209
-          ],
-          [
-            -73.6410079055759,
-            45.4584725211565
-          ],
-          [
-            -73.6408517473394,
-            45.458413027044
-          ],
-          [
-            -73.6408116759182,
-            45.4584611281071
-          ],
-          [
-            -73.6407504451182,
-            45.4584372457533
-          ],
-          [
-            -73.6407102295275,
-            45.4584887179595
-          ],
-          [
-            -73.6407733946252,
-            45.4585134741824
-          ],
-          [
-            -73.6407529936233,
-            45.4585395360268
-          ],
-          [
-            -73.6408772869513,
-            45.4585890444863
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'RF2',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.640685034277,
-            45.4585208608007
-          ],
-          [
-            -73.6407957736703,
-            45.4583791494107
-          ],
-          [
-            -73.6405590484041,
-            45.458288067748
-          ],
-          [
-            -73.640564372429,
-            45.4582807370996
-          ],
-          [
-            -73.6402606533331,
-            45.4581655166129
-          ],
-          [
-            -73.6402556996114,
-            45.4581718691469
-          ],
-          [
-            -73.6400146981164,
-            45.4580776515112
-          ],
-          [
-            -73.6399947518581,
-            45.4581036322558
-          ],
-          [
-            -73.6399668476527,
-            45.4580924575034
-          ],
-          [
-            -73.6399436219669,
-            45.4581219391978
-          ],
-          [
-            -73.6399074228682,
-            45.4581082196506
-          ],
-          [
-            -73.6399223691219,
-            45.4580900595336
-          ],
-          [
-            -73.6399135855623,
-            45.4580862631194
-          ],
-          [
-            -73.6400076738254,
-            45.4579642333517
-          ],
-          [
-            -73.6400664064559,
-            45.4579871594348
-          ],
-          [
-            -73.6401266871755,
-            45.4579119371099
-          ],
-          [
-            -73.6398312888425,
-            45.4577985269392
-          ],
-          [
-            -73.6397706190755,
-            45.4578764941739
-          ],
-          [
-            -73.6398149533159,
-            45.4578930300438
-          ],
-          [
-            -73.6398066969935,
-            45.4579036174965
-          ],
-          [
-            -73.6398351180342,
-            45.4579163430144
-          ],
-          [
-            -73.6397535448525,
-            45.4580219205454
-          ],
-          [
-            -73.639741129547,
-            45.4580208358951
-          ],
-          [
-            -73.6396715170233,
-            45.4579983776984
-          ],
-          [
-            -73.6396309215869,
-            45.4580654273612
-          ],
-          [
-            -73.6396882783852,
-            45.4580865293883
-          ],
-          [
-            -73.6396887773383,
-            45.4580898474012
-          ],
-          [
-            -73.6396951868686,
-            45.4580959502411
-          ],
-          [
-            -73.6396166169914,
-            45.4582019689471
-          ],
-          [
-            -73.6395811225177,
-            45.4581898275843
-          ],
-          [
-            -73.6395737199856,
-            45.4581996554648
-          ],
-          [
-            -73.6395088037371,
-            45.4581730191629
-          ],
-          [
-            -73.6394429805971,
-            45.4582552453557
-          ],
-          [
-            -73.6397634323295,
-            45.4583770539877
-          ],
-          [
-            -73.639823318343,
-            45.4582976433774
-          ],
-          [
-            -73.6397656937339,
-            45.458276649559
-          ],
-          [
-            -73.6398615938007,
-            45.4581522303785
-          ],
-          [
-            -73.6398702925016,
-            45.4581551294048
-          ],
-          [
-            -73.6398811620684,
-            45.4581418281119
-          ],
-          [
-            -73.63991569451,
-            45.4581545656232
-          ],
-          [
-            -73.6399458518892,
-            45.4581660710766
-          ],
-          [
-            -73.6399018618318,
-            45.4582187587328
-          ],
-          [
-            -73.6401444564348,
-            45.4583151280049
-          ],
-          [
-            -73.6401414600105,
-            45.4583216203185
-          ],
-          [
-            -73.640445805528,
-            45.45843855574
-          ],
-          [
-            -73.6404494963462,
-            45.4584315753495
-          ],
-          [
-            -73.640685034277,
-            45.4585208608007
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'TA-TB',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.64089299778541,
-            45.46005310876819
-          ],
-          [
-            -73.64097081766164,
-            45.459958181111034
-          ],
-          [
-            -73.64087044130008,
-            45.45991862787304
-          ],
-          [
-            -73.64079713272129,
-            45.46001118240605
-          ],
-          [
-            -73.64089299778541,
-            45.46005310876819
-          ]
-        ]
-      ]
-    },
-  },
-  {
-    type: 'Feature',
-    properties: {
-      height: 1,
-      color: '#912338',
-      id: 'PY',
-    },
-
-    geometry: {
-      type: 'Polygon',
-      coordinates: [
-        [
-          [
-            -73.64055934834015,
-            45.45920946153353
-          ],
-          [
-            -73.64062515971617,
-            45.45913134250955
-          ],
-          [
-            -73.64056567635693,
-            45.45910914958574
-          ],
-          [
-            -73.64080993780989,
-            45.458796672285104
-          ],
-          [
-            -73.6404492408458,
-            45.458651973403704
-          ],
-          [
-            -73.6404201319684,
-            45.45869014553736
-          ],
-          [
-            -73.64040367912459,
-            45.45868748236603
-          ],
-          [
-            -73.64022649465055,
-            45.45891562691625
-          ],
-          [
-            -73.64023788508125,
-            45.45891917779713
-          ],
-          [
-            -73.64013916801758,
-            45.45904434619709
-          ],
-          [
-            -73.64055934834015,
-            45.45920946153353
-          ]
-        ]
-      ]
-    },
-  }
-];
 
 const fixPolygonCoordinates = (coordinates: number[][][]): number[][][] => {
   return coordinates.map((ring) => {
